[Closes #16] Added Widget example/test class
Unlike originally thought, this will not be a
virtual class to inherit from, but rather an
example class to copy/paste in order to make new
Widgets. This is just the way React prefers it.
As long as the Prop interface is the same, all
components should be fairly interchangeable.