[Closes #30]
The webapp works on Heroku. Only the
quantum-hub/api folder is pushed to Heroku.
To make this work, yarn build in package.json
has been redefined to build and move the build
to api/
api.py has also been changed, it is no longer
called with flask, but instead simply with
python.
NOTE: if you are pulling, please remove
package.json from your .gitignore, or else
the new build macro won't be found.

NOTE II: Instructions on deployment to Heroku to
come soon.