[Closes #35] Update sudoku.py to work with the api.py Endpoint
What:
- Pass in qpu_sampler to main() for use when sampling
from the KerberosSampler to get the solution to the sudoku problem
- Return solved sudoku board

Why:
- We will not rely on auto-detected configuration files
when creating a client but will do so based on both pre-set
parameters and those entered by the user. Therefore, we will
explicitly pass in a qpu_sampler constructed in the calling-function
of main() using the aforementioned parameters.
- Return value is needed to pass to the client-side to display on the UI