[Closes #53] Update sudoku.py to handle JSON response
PS1:
What: sudoku.py specifies a JSON response from its xhr and
outputs to the solved sudoku board received to the console.

TODO: Set the sudoku grid to reflect the solved board state.

PS2:
What:
- Set the sudoku grid to reflect the solved board state.
- Changed console output from
'Sending this grid in for solving:' to
'Sending in this grid:' so it fits on one line