[Closes #74] Add Reset Button for Sudoku App
What:
Add Reset Button that is enabled when:
1. The Sudoku Grid is not empty
2. The Sudoku Grid is not loading

The Reset Button clears the Sudoku Grid and becomes disabled.
The left-hand console is not reset.

- Includes minor StandardJS formatting changes

NOTE: The file sudoku.js is displaying some code smells (i.e. bloaters)
and (since the file is still young) a refactor should be made.
Issue #75 has been opened to track progress.