[Closes #79] Change Description Card Component
What: Switch to using Material-Kit-React's
Card component instead of Material-UI's
Complex Interaction card.
Why: The description was too long to fit in the old component.
Although it did expand, it would extend the shell body so far down
that the whitespace to its left made the design look incomplete.
The new component will click to open a Dialog over the app shell
which the user can close.

NOTE: Description Card width of 314.9765px is chosen because:
- the Grid Container's max width is 1140px
- the Sudoku Game's div is 462.047px + 2 * 8px (margin of 8px on each side)
- this leaves ((1140px - (462.047px + 2 * 8px)) / 2) - 2 * 8px
(margin of 8px on each side) = 314.9765px equal width for the
Console and Description Card.