[#87] Invalid Sudoku Board is Solved
What: Entering an invalid Sudoku board (i.e. with all 4's
in the first column of the first grid) highlights:
- the invalid input
- the invalid row
- the invalid column
- the invalid subgrid
but does not prevent the user from submitting the board
to be solved.

TODOS:
- Consider setting focussed grid square to nothing instead
of the top-left grid when board is loading during the xhr
request
- Add documentation
- #75
- Add getSudokuValidationResults() Function Citations

PS2: All entries in a row, column or subgrid with duplicates
will now be highlighted. Before, all duplicate entries except
the first in the row, column or subgrid were highlighted