[Closes #87] Invalid Sudoku Board is Solved
What: Entering an invalid Sudoku board (i.e. with all 4's
in the first column of the first grid) highlights:
- the invalid input
- the invalid row
- the invalid column
- the invalid subgrid
but does not prevent the user from submitting the board
to be solved.

PS2: All entries in a row, column or subgrid with duplicates
will now be highlighted. Before, all duplicate entries except
the first in the row, column or subgrid were highlighted

PS3:
- [#75] Created SudokuValidationUtils.js
- Sudoku grid will have 0.65 opacity when loading to look
disabled
- Upon Sudoku Solve request completing (or failing) the
Sudoku Grid's state is revalidated instead of reset